requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.11.2
argparse==1.2.1
future==0.15.2
httplib2==0.9.2
itsdangerous==0.24
nose==1.3.7
omnijson==0.1.2
python-dateutil==1.5
python-telegram-bot==2.9.2
six==1.10.0
wsgiref==0.1.2
flask-mongoengine
